% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airship.R
\name{airship}
\alias{airship}
\title{Runs the app "AIRSHIP".}
\usage{
airship(dfData = NULL, cLastInputVar = NULL, cReplicationVar = NULL)
}
\arguments{
\item{dfData}{Dataset that should be plotted by Airship; can be NULL if
upload should be done within the app.}

\item{cLastInputVar}{Optional and only useful in combination with dfData.
Character name of last input variable.}

\item{cReplicationVar}{Optional and only useful in combination with dfData.
Character name of simulation replication variable.}
}
\value{
No return value
}
\description{
Runs the app "AIRSHIP".
}
\examples{

\dontrun{
airship()
airship(data)
}

# See Vignette.

}
