% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airship.R
\name{airship}
\alias{airship}
\title{Runs the app "AIRSHIP".}
\usage{
airship(...)
}
\arguments{
\item{...}{Further arguments to be passed to Shiny's "shinyApp()" function}
}
\value{
No return value
}
\description{
Runs the app "AIRSHIP".
}
\examples{
\dontrun{
airship()
}

}
