% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airship.R
\name{airship}
\alias{airship}
\title{Runs the Shiny app "AIRSHIP".}
\usage{
airship(
  dfData = NULL,
  cLastInputVar = NULL,
  cReplicationVar = NULL,
  bIsFacts = FALSE
)
}
\arguments{
\item{dfData}{Dataset that should be plotted by Airship; can be NULL if
upload should be done within the app.}

\item{cLastInputVar}{Optional and only useful in combination with dfData.
Character name of last input variable.}

\item{cReplicationVar}{Optional and only useful in combination with dfData.
Character name of simulation replication variable.}

\item{bIsFacts}{Boolean variable; is the supplied dfData a FACTS aggregated simulation file.}
}
\value{
No return value
}
\description{
Runs the Shiny app "AIRSHIP".
}
\examples{

if(interactive()){
airship()
}

# See Vignette.

}
