% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleData2.R
\docType{data}
\name{ExampleData2}
\alias{ExampleData2}
\title{Example Data 2}
\format{
An object of class \code{data.frame} with 10080 rows and 18 columns.
}
\source{
https://github.com/el-meyer/airship/blob/master/data/ExampleDataNASH.csv
}
\usage{
ExampleData2
}
\description{
Simulated dataset from Meyer et al. (2022) https://doi.org/10.1002/pst.2194.
}
\keyword{datasets}
